Added Routes to Home Page and Interest List Page
I created a page to display all entries from an Interest List. It includes the following information: name, phone number, event type, date, number of spaces, whether alcohol will be served, whether the client was quoted, the tour date, status of the lead, and the date of the last update on the lead.